refactor(query): migrate to common-users
