+++ query/Cargo.toml
@@ -56,6 +56,7 @@ common-metrics = { path = "../common/metrics" }
 common-planners = { path = "../common/planners" }
 common-streams = { path = "../common/streams" }
 common-tracing = { path = "../common/tracing" }
+common-users = { path = "../common/users" }
 
 # Github dependencies
 async-trait = { git = "https://github.com/datafuse-extras/async-trait", rev = "f0b0fd5" }
@@ -86,7 +87,6 @@ futures-util = "0.3.21"
 headers = "0.3.7"
 http = "0.2.8"
 itertools = "0.10.3"
-jwtk = "0.2.3"
 lz4 = "1.23.3"
 metrics = "0.19.0"
 naive-cityhash = "0.2.0"
