+++ query/tests/it/auth.rs
@@ -17,9 +17,9 @@ use base64::URL_SAFE_NO_PAD;
 use common_base::base::tokio;
 use common_exception::Result;
 use common_meta_types::UserIdentity;
-use databend_query::users::auth::jwt::CustomClaims;
-use databend_query::users::auth::jwt::EnsureUser;
-use databend_query::users::Credential;
+use common_users::CustomClaims;
+use common_users::EnsureUser;
+use databend_query::auth::Credential;
 use jwt_simple::prelude::*;
 use wiremock::matchers::method;
 use wiremock::matchers::path;
