+++ query/tests/it/main.rs
@@ -11,7 +11,9 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+
 mod api;
+mod auth;
 mod catalogs;
 mod clusters;
 mod common;
@@ -27,5 +29,4 @@ mod sql;
 mod storages;
 mod table_functions;
 mod tests;
-mod users;
 mod version;
